requirements.txt
flask
flask-login
flask-session
redis
Flask-SQLAlchemy

